backend/requirements.txt
django==4.2
psycopg2-binary==2.9.10
djangorestframework==3.14.0